Mail Drafts vs. Oracle
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@33762 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/Mail/classes/class.ilMailFormGUI.php
@@ -215,8 +215,8 @@ class ilMailFormGUI
 				ilUtil::securePlainString($_POST["m_email"]),
 				ilUtil::securePlainString($_POST["m_subject"]),
 				ilUtil::securePlainString($_POST["m_message"]),
-				$_SESSION["draft"],
-				ilUtil::securePlainString($_POST['use_placeholders'])
+				(int)$_SESSION["draft"],
+				(int)ilUtil::securePlainString($_POST['use_placeholders'])
 			);
 			#session_unregister("draft");
 			#ilUtil::sendInfo($this->lng->txt("mail_saved"),true);
